Fix memory leak and compiler warnings
